Change type of `WriterOptions.DictSize`

+++ reader.go
@@ -45,8 +45,8 @@ func NewReader(r io.Reader) (*Reader, error) {
 		return nil, ErrUnknownVersion
 	}
 
-	dictSize := 1 << (int(header[5] & 0x1f))
-	dictSize -= (dictSize / 16) * int((header[5]>>5)&0x07)
+	dictSize := uint32(1 << (header[5] & 0x1f))
+	dictSize -= (dictSize / 16) * uint32((header[5]>>5)&0x07)
 
 	switch {
 	case dictSize < MinDictSize:
@@ -63,7 +63,7 @@ func NewReader(r io.Reader) (*Reader, error) {
 
 	var lzmaHeader [lzma.HeaderLen]byte
 	lzmaHeader[0] = lzma.Properties{LC: 3, LP: 0, PB: 2}.Code()
-	binary.LittleEndian.PutUint32(lzmaHeader[1:5], uint32(dictSize))
+	binary.LittleEndian.PutUint32(lzmaHeader[1:5], dictSize)
 	copy(lzmaHeader[5:], rb[len(rb)-16:len(rb)-8])
 
 	z.trailer.memberSize = uint64(headerSize + len(rb))
